Fix ubuntu warning

+++ interface/src/octree/SafeLanding.cpp
@@ -170,7 +170,7 @@ bool SafeLanding::isEntityPhysicsReady(const EntityItemPointer& entity) {
             entity->getAABox(hasAABox);
             if (hasAABox && downloadedCollisionTypes.count(modelEntity->getShapeType()) != 0) {
                 auto space = _entityTreeRenderer->getWorkloadSpace();
-                uint8_t region = space ? space->getRegion(entity->getSpaceIndex()) : workload::Region::INVALID;
+                uint8_t region = space ? space->getRegion(entity->getSpaceIndex()) : (uint8_t)workload::Region::INVALID;
                 bool shouldBePhysical = region < workload::Region::R3 && entity->shouldBePhysical();
                 return (!shouldBePhysical || entity->isInPhysicsSimulation() || modelEntity->computeShapeFailedToLoad());
             }
